Extract SSHConfig from config.py and move to ssh_config.py (more user-friendly)
Rollback API version to 0.2
